ZON-6127: Be more precise in how we specify polymorphic items
Wherever an array can contain different items (that share a common
base, but then offer different fields, based on a distinctive `type`
field), the statement that we want to make is that each item
distinctly belongs to _exactly one_ of the possible subtypes.

OpenAPI expresses this as `oneOf`, whereas `anyOf` means "matches _at
least_ one of the possibilities", which is still true, but not as
precise as it should be.