ZON-6277: provide information about user
this essentially is a manual and complete version of the auto-generated
spec from the profile server at
https://docs.zeit.de/comment/referenz/rest.html

since that spec is auto-generated (and thus more or less useless, since
it says nothing about the actual content) we recreate it here.

note that we (indirectly) signify the login status via return code. this
may or may not be a desirable behaviour but seems like a good, sparse
initial approach.