MAINT: make the generic 400 responses valid for all resources
otherwise we have the absurd situation that openapi **request**
validation  detects an invalid
request, it produces a 400 with a (usually) helpful error message but if
400 is not a valid response, openapi's **response** validation produces
a 500 (without any useful information whatsoever)

this PR simply sprinkles 400 responses over those resources that haven't
declared them yet. henceforth we need to add them to begin with when
configuring new resources.