Changed headerimage and added 2 more things
Thanks @thomaspuppe for giving me the tip that `headerimage` was a legacy teaser layout. I have changed `headerimage` to `zon-teaser-panorama` and also included podcast-episode-list and zon-teaser-podcast-small-plus. Sound weird but there is a very compact podcast teaser in the new podcast tab of the app (zon-teaser-podcast-compact) and a version of the same teaser with teaser text (zon-teaser-small-compact-plus)